refactor: 에러 메세지 ENUM 리팩토링
