+++ src/main/java/lotto/domain/lotto/LottoLine.java
@@ -1,24 +1,24 @@
 package lotto.domain.lotto;
 
+import static lotto.view.messages.ErrorMessages.LOTTO_LINE_NUMBER_COUNT_DUPLICATE_ERROR;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import lotto.domain.lotto.utils.Rank;
-import lotto.view.ErrorMessages;
 
 public class LottoLine {
 
     private final Set<LottoNumber> value;
 
     public LottoLine(List<LottoNumber> lottoNumbers) {
-        Set<LottoNumber> unDuplicatedLottoNumber = new HashSet(lottoNumbers);
-        if (unDuplicatedLottoNumber.size() != 6) {
-            throw new IllegalArgumentException(
-                ErrorMessages.ERROR_LOTTO_NUMBER_DUPLICATED.getMessage());
+        Set<LottoNumber> set = new HashSet<>(lottoNumbers);
+        if (set.size() != 6) {
+            throw new IllegalArgumentException(LOTTO_LINE_NUMBER_COUNT_DUPLICATE_ERROR.getMessage());
         }
-        value = unDuplicatedLottoNumber;
+        value = set;
     }
 
     public Rank matchLottoNumbers(List<LottoNumber> lottoNumbers, LottoNumber bonusNumber,
